[MIG] stock_demand_estimate_matrix: Migration to 14.0
