+++ stock_demand_estimate_matrix/__manifest__.py
@@ -3,12 +3,13 @@
 {
     "name": "Stock Demand Estimate Matrix",
     "summary": "Allows to create demand estimates.",
-    "version": "13.0.1.0.0",
+    "version": "14.0.1.0.0",
     "author": "ForgeFlow, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/stock-logistics-warehouse",
-    "category": "Warehouse Management",
+    "category": "Warehouse",
     "depends": ["stock_demand_estimate", "web_widget_x2many_2d_matrix", "date_range"],
     "data": [
+        "security/ir.model.access.csv",
         "views/stock_demand_estimate_view.xml",
         "views/date_range.xml",
         "wizards/stock_demand_estimate_wizard_view.xml",
