+++ stock_demand_estimate_matrix/wizards/__init__.py
@@ -1,3 +1,5 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 from . import stock_demand_estimate_wizard
+from . import stock_demand_estimate_sheet
+from . import stock_demand_estimate_sheet_line
